Explicitly disable input length restriction for EditPayloadFragment.
Works around Samsung's odd default 9000 character restriction for TextViews.
Today I was in a meeting taking notes and input stopped when I hit 9000
chars. So I saved the note and made a new one, then later merged the two
on my desktop. When I synced that back to MobileOrg, I got a crash when I
tried to view the 12K note. This patch seems to fix the crash and also
the input restriction.

see also:

http://www.elmoensio.com/2012/03/11/gingerbread-9000-character-limit/

http://stackoverflow.com/a/3681475/589306