Split queries into Agenda and TODO list types.
The eventual goal is to allow for construction of most queries provided
by Emacs Org Mode's variable org-agenda-custom-commands. Toward this
goal, we have added a few options and views for them to the Agenda entry
settings.

Don't convert all day events to UTC. Doing so makes them come after
events starting at midnight (00:00) of the same day in timezones east of
UTC, and before such events in timezones west of UTC. By default, Emacs
Org Mode converts dates to days since the epoch, but it can use seconds,
(org-agenda-todo-ignore-time-comparison-use-seconds), so it is important
to know what time zone it uses. The last element of decode-time's return
value is "an integer indicating the number of seconds east of
Greenwich," which (decode-time (current-time)) gives as -25200
(GMT-7:00, MST) on my system. Therefore Emacs calculates times in the
local time zone. We should emulate this behavior so the same entries
appear in our agenda queries as in Org Mode's.